% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DiegoTestPackg.R
\name{cuadrado}
\alias{cuadrado}
\title{Cuadrado de un número}
\usage{
cuadrado(x)
}
\arguments{
\item{Valor}{numérico pasado a la función}
}
\value{
El cuadrado del número pasado como parámetro
}
\description{
Eleva cualquier nujmero al cuadrado
}

