\name{DiegoTestPackg-package}
\alias{DiegoTestPackg-package}
\alias{DiegoTestPackg}
\docType{package}
\title{
Paquete de prueba Diego Paniagua
}
\description{
~~ Paquete de prueba que unicamente incluye la función cuadrado
}
\details{
\tabular{ll}{
Package: \tab DiegoTestPackg\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-01-02\cr
License: \tab Free\cr
}
~~ Paquete de prueba que unicamente incluye la función cuadrado
}
\author{
Diego Paniagua <dpansan@gmail.com>

Maintainer: Diego Paniagua <dpansan@gmail.com>
}
\references{
http://git-scm.com/book/es/v1/Fundamentos-de-Git-Guardando-cambios-en-el-repositorio
http://r-pkgs.had.co.nz/man.html
}

\examples{
}
